Fix HTML help build with latest tools.
Updated environment wrapper script to be more
compatible with typical 'python3' environments
so that an external 'textile' module is more
likely to be found successfully.

Improved error-handling in the HTML generator
to make Textile import issues more obvious in
the HTML output itself.

Fixed a couple of Textile input files with
syntax errors that produced wrong output on
newer parsers (and happened to work with the
older ones).

Fixed date substitution in index file to use
a different placeholder that will not be
reinterpreted by newer Textile parsers.